5. Create a CatViewModel. This class should:
9

- Store the URL used to access the API in a String property named
urlString, initialized to:
https://catfact.ninja/breeds
- Have a property named breeds which is an array of CatBreed
structs, and which will hold data returned by the data array by
the above URL.
- Create a function called getData that will do whatever is
necessary to access the API at the URL in urlString, and return
the data, assigning this to the breeds properties of the
CatViewModel class.
Hint: you may want to simply access a single variable to make sure that you can get data, then
expand your work to make sure you can get the array of data that you need.
- Do whatever else is necessary for this to act like a ViewModel that can access data from an API,
parse the data into the breeds