13.Scrolling and loading paged JSON: If you look at the JSON at this URL: https://catfact.ninja/breeds and scroll to the end of the cat breeds, you should notice a key value in the JSON that's named: next_page_url
This will be either a String containing the URL to load the next page of 25 cat breeds, or it will be null if
you have loaded the last available page. The implementation is very similar to what you've done for
your CatchEmAll app, so there should be no surprises on how to get this to work.
- After you've downloaded and parsed JSON data, update urlString with the value from
next_page_url. If next_page_url returns null, you can set urlString to an empty String. This can
be how you can tell if you don’t have any more pages to load.
- Using techniques learned in this course, create an additional function in your CatViewModel,
named loadNextIfNeeded. The function should accept a CatBreed parameter named catBreed,
and load the next page of data, using the updated urlString if the catBreed passed into this
function is the last cat breed that has been downloaded to the app & there are still pages to load.

13
- Modify your code in ListView so that you can properly call loadNextIfNeeded as each cat bread
displays in the List, passing in the current catBreed being displayed.
If everything works properly, you should find the next page of data loads as you scroll, but you
won't be able to scroll past the last breed on the last page available (as of the creation of this
exam, the last breed on the last page was named "York Chocolate", and this was the 98th breed).